requirements.txt
PyGithub~=2.1.1
pandas~=2.1.4
